// src/generated/prisma/enums.ts
export const ConsultationMode = {
  IN_PERSON: 'IN_PERSON',
  VIDEO: 'VIDEO',
  PHONE: 'PHONE'
} as const

export type ConsultationMode = (typeof ConsultationMode)[keyof typeof ConsultationMode]